Remove extra types file
