+++ src/views/open-api-docs-view/components/open-api-overview/components/overview-blurb/index.tsx
@@ -1,10 +1,17 @@
 import InlineLink from 'components/inline-link'
-import { OpenApiOverviewProps } from '../../types'
+import { OpenApiOverviewProps } from '../../'
 import s from './overview-blurb.module.css'
 
 /**
  * TODO extract out of inline content, figure out where we want to keep this data
  * ideally somewhere with the open API data.
+ *
+ * Note from Zach:
+ * 📌 For the future state where this content lives in the spec file itself,
+ * one possibility might be the description property of the info section of the schema.
+ * CommonMark is specifically mentioned as something that "MAY" be used for that field.
+ * I think this content could be represented in markdown, and I think "should be representable in markdown"
+ * might be a good guardrail for content in this area in the meantime 👍
  * */
 
 export default function OverviewBlurb({
