--- src/views/open-api-docs-view/components/open-api-overview/types.ts
@@ -1,10 +1,0 @@
-export interface OpenApiOverviewProps {
-	title: string
-	badgeText: string
-	description: string
-	status?: {
-		text: string
-		href: string
-	}
-	className?: string
-}
